refactor: apply helper functions to reduce code duplication
- Replace manual item.item || item patterns with getActualItem()
- Use TITLE_FIELDS constant instead of hardcoded display field arrays
- Use METADATA_FIELDS constant instead of hardcoded system field arrays
- Replace manual collection parsing with parseAllowedCollections()
- Update utils to use logger-wrapper instead of direct logger import
- Remove redundant wrapper functions in useUIHelpers
